moved sample read event handler into it's own file / class
